requirements.txt
aiofiles==0.8.0
anyio==3.5.0
azure-ai-formrecognizer==3.1.2
azure-common==1.1.28
azure-core==1.23.1
azure-identity==1.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==36.0.2
Flask==2.1.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.6.0
frontend==0.0.3
idna==3.3
importlib-metadata==4.11.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
numpy==1.22.3
oauthlib==3.2.0
pandas==1.4.2
portalocker==2.4.0
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
starlette==0.18.0
threadpoolctl==3.1.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.4
Werkzeug==2.1.2
wincertstore==0.2
XlsxWriter==3.0.3
zipp==3.8.0
azure-storage
azure-storage-blob
